Slim Elastic.Documentation by extracting heavy deps to new Tooling project
Create Elastic.Documentation.Tooling to hold infrastructure types (ExternalCommandExecutor,
GitCheckoutInformationFactory, Exporter, Paths, FileSystemFactory) that dragged YamlDotNet,
Proc, IniFileParser, and Nullean.Argh into the domain library. Move YAML converters
(SemVersionConverter, ApplicableToYamlConverter) into Elastic.Markdown where they are consumed.
Swap YamlMember/YamlSerializable attributes on AppliesTo types for JsonPropertyName equivalents
so the domain library no longer depends on YamlDotNet at all.
