+++ docs-builder.slnx
@@ -53,6 +53,7 @@
     <Project Path="src/Elastic.Documentation.Site/Elastic.Documentation.Site.csproj" />
     <Project Path="src/Elastic.Documentation.Svg/Elastic.Documentation.Svg.csproj" />
     <Project Path="src/Elastic.Documentation/Elastic.Documentation.csproj" />
+    <Project Path="src/Elastic.Documentation.Tooling/Elastic.Documentation.Tooling.csproj" />
     <Project Path="src/Elastic.Markdown/Elastic.Markdown.csproj" />
   </Folder>
   <Folder Name="/src/api/">
