+++ src/Elastic.Documentation.Tooling/Elastic.Documentation.Tooling.csproj
@@ -6,17 +6,20 @@
     <Nullable>enable</Nullable>
     <RootNamespace>Elastic.Documentation</RootNamespace>
     <IsAotCompatible>true</IsAotCompatible>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <NoWarn>$(NoWarn);CS1591;CS1573;CS1572;CS1571;CS1570;CS1574;CS0419</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Elastic.Mapping" />
-    <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" />
-    <PackageReference Include="Microsoft.Extensions.Logging"/>
-    <PackageReference Include="NetEscapades.EnumGenerators"/>
+    <ProjectReference Include="..\Elastic.Documentation\Elastic.Documentation.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging" />
+    <PackageReference Include="Nullean.Argh.Interfaces" />
     <PackageReference Include="Nullean.ScopedFileSystem" />
+    <PackageReference Include="Proc" />
+    <PackageReference Include="SoftCircuits.IniFileParser" />
     <PackageReference Include="System.IO.Abstractions" />
+    <PackageReference Include="System.IO.Abstractions.TestingHelpers" />
   </ItemGroup>
 
 </Project>
