+++ src/Elastic.Documentation.Tooling/ExternalCommands/ExternalCommandExecutor.cs
