+++ src/Elastic.Documentation.Tooling/ExternalCommands/NoopConsoleWriter.cs
