+++ src/Elastic.Documentation/AppliesTo/ProductLifecycle.cs
@@ -2,40 +2,39 @@
 // Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
 // See the LICENSE file in the project root for more information
 
-using YamlDotNet.Serialization;
+using System.Text.Json.Serialization;
 
 namespace Elastic.Documentation.AppliesTo;
 
-[YamlSerializable]
 public enum ProductLifecycle
 {
 	// technical preview (exists in current docs system per https://github.com/elastic/docs?tab=readme-ov-file#beta-dev-and-preview-experimental)
-	[YamlMember(Alias = "preview")]
+	[JsonStringEnumMemberName("preview")]
 	TechnicalPreview,
 	// beta (ditto)
-	[YamlMember(Alias = "beta")]
+	[JsonStringEnumMemberName("beta")]
 	Beta,
 	// ga (replaces "added" in the current docs system since it was not entirely clear how/if that overlapped with beta/preview states)
-	[YamlMember(Alias = "ga")]
+	[JsonStringEnumMemberName("ga")]
 	GenerallyAvailable,
 	// deprecated (exists in current docs system per https://github.com/elastic/docs?tab=readme-ov-file#additions-and-deprecations)
-	[YamlMember(Alias = "deprecated")]
+	[JsonStringEnumMemberName("deprecated")]
 	Deprecated,
 	// removed content
-	[YamlMember(Alias = "removed")]
+	[JsonStringEnumMemberName("removed")]
 	Removed,
 	// unavailable (for content that doesn't exist in a specific context and is never coming or not coming anytime soon)
-	[YamlMember(Alias = "unavailable")]
+	[JsonStringEnumMemberName("unavailable")]
 	Unavailable,
 
 	// TODO remove these enum members in a future version when docs have been cleaned up
 	// discontinued (historically we've immediately removed content when the feature ceases to be supported, but this might not be the case with pages that contain information that spans versions)
-	[YamlMember(Alias = "discontinued")]
+	[JsonStringEnumMemberName("discontinued")]
 	Discontinued,
 	// coming (ditto)
-	[YamlMember(Alias = "planned")]
+	[JsonStringEnumMemberName("planned")]
 	Planned,
 	// dev (ditto, though it's uncertain whether it's ever used or still needed)
-	[YamlMember(Alias = "development")]
+	[JsonStringEnumMemberName("development")]
 	Development,
 }
