+++ src/Elastic.Markdown/Myst/YamlSerialization.cs
@@ -38,7 +38,6 @@ public static class YamlSerialization
 [YamlSerializable(typeof(Setting))]
 [YamlSerializable(typeof(AllowedValue))]
 [YamlSerializable(typeof(SettingMutability))]
-[YamlSerializable(typeof(ApplicableTo))]
 [YamlSerializable(typeof(ContributorEntry))]
 [YamlSerializable(typeof(ChangelogDirectiveConfigYaml))]
 [YamlSerializable(typeof(ChangelogDirectiveBundleConfigYaml))]
