+++ tests/Elastic.Documentation.Configuration.Tests/ConfigurationFileExcludeTests.cs
@@ -93,6 +93,6 @@ public class ConfigurationFileExcludeTests
 		public IFileInfo ConfigurationPath => configurationPath;
 		public BuildType BuildType => BuildType.Isolated;
 		public IDirectoryInfo DocumentationSourceDirectory => documentationSourceDirectory;
-		public GitCheckoutInformation Git => GitCheckoutInformation.Create(documentationSourceDirectory, fileSystem);
+		public GitCheckoutInformation Git => GitCheckoutInformationFactory.Create(documentationSourceDirectory, fileSystem);
 	}
 }
